# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Smart Investment Advisor",
    "description": "AI 智能投资助手后端 API，集成多源数据与 LLM 分析能力",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/portfolio/search": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Search Stocks",
        "description": "股票搜索接口\n- 同时支持本地数据库模糊搜索和远程 API 实时搜索\n- 如果本地没搜到且开启了 remote=True，会尝试从第三方源抓取并同步到本地",
        "operationId": "search_stocks_api_portfolio_search_get",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Query"
            }
          },
          {
            "name": "remote",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Remote"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  },
                  "title": "Response Search Stocks Api Portfolio Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/summary": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Get Portfolio Summary",
        "description": "获取投资组合汇总数据 (Get Portfolio Summary)",
        "operationId": "get_portfolio_summary_api_portfolio_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/portfolio/": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Get Portfolio",
        "description": "获取当前用户的投资组合列表 (Fetch User Portfolio)\n\n逻辑 (Logic)：\n1. 联表查询：Portfolio (持仓) -> MarketDataCache (行情缓存) -> Stock (基础资料)\n2. 若 refresh=True，则同步/异步触发全量实时刷新\n3. 循环计算盈亏比例、市值、未实现盈亏等前端展示字段",
        "operationId": "get_portfolio_api_portfolio__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Refresh"
            }
          },
          {
            "name": "price_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Price Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortfolioItem"
                  },
                  "title": "Response Get Portfolio Api Portfolio  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "portfolio"
        ],
        "summary": "Add Portfolio Item",
        "description": "添加或更新自选股/持仓",
        "operationId": "add_portfolio_item_api_portfolio__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/{ticker}": {
      "delete": {
        "tags": [
          "portfolio"
        ],
        "summary": "Delete Portfolio Item",
        "description": "从自选股中删除",
        "operationId": "delete_portfolio_item_api_portfolio__ticker__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/{ticker}/refresh": {
      "post": {
        "tags": [
          "portfolio"
        ],
        "summary": "Refresh Stock Data",
        "description": "针对单个股票的手动强制刷新接口 (支持极速模式)",
        "operationId": "refresh_stock_data_api_portfolio__ticker__refresh_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "price_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Price Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/{ticker}/news": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Get Stock News",
        "description": "获取该股票在数据库中存储的最新新闻",
        "operationId": "get_stock_news_api_portfolio__ticker__news_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/reorder": {
      "patch": {
        "tags": [
          "portfolio"
        ],
        "summary": "Reorder Portfolio",
        "description": "批量更新排序权重 (Batch update sort_order)",
        "operationId": "reorder_portfolio_api_portfolio_reorder_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array",
                "title": "Orders"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/stocks/{ticker}/history": {
      "get": {
        "tags": [
          "stocks"
        ],
        "summary": "Get Stock History",
        "description": "接口：获取股票的历史行情数据，专为 K 线图打造。",
        "operationId": "get_stock_history_api_stocks__ticker__history_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1y",
              "title": "Period"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1d",
              "title": "Interval"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OHLCVItem"
                  },
                  "title": "Response Get Stock History Api Stocks  Ticker  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stocks/refresh_all": {
      "post": {
        "tags": [
          "stocks"
        ],
        "summary": "Refresh All Stocks",
        "description": "接口：一键同步当前数据库中所有活跃股票的最新行情。",
        "operationId": "refresh_all_stocks_api_stocks_refresh_all_post",
        "parameters": [
          {
            "name": "price_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Price Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/portfolio": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Portfolio Analysis",
        "description": "获取最近的一次全量持仓健康分析 (非实时生成)",
        "operationId": "get_portfolio_analysis_api_analysis_portfolio_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioAnalysisResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyze Portfolio",
        "description": "全量持仓健康分析接口 (Portfolio Health Check)",
        "operationId": "analyze_portfolio_api_analysis_portfolio_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioAnalysisResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OHLCVItem": {
        "properties": {
          "time": {
            "type": "string",
            "title": "Time"
          },
          "open": {
            "type": "number",
            "title": "Open"
          },
          "high": {
            "type": "number",
            "title": "High"
          },
          "low": {
            "type": "number",
            "title": "Low"
          },
          "close": {
            "type": "number",
            "title": "Close"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "rsi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rsi"
          },
          "macd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Macd"
          },
          "macd_signal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Macd Signal"
          },
          "macd_hist": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Macd Hist"
          },
          "bb_upper": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bb Upper"
          },
          "bb_middle": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bb Middle"
          },
          "bb_lower": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bb Lower"
          }
        },
        "type": "object",
        "required": [
          "time",
          "open",
          "high",
          "low",
          "close"
        ],
        "title": "OHLCVItem"
      },
...
      "PortfolioAnalysisResponse": {
        "properties": {
          "health_score": {
            "type": "integer",
            "title": "Health Score"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "diversification_analysis": {
            "type": "string",
            "title": "Diversification Analysis"
          },
          "strategic_advice": {
            "type": "string",
            "title": "Strategic Advice"
          },
          "top_risks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Top Risks"
          },
          "top_opportunities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Top Opportunities"
          },
          "detailed_report": {
            "type": "string",
            "title": "Detailed Report"
          },
          "model_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Used"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "health_score",
          "risk_level",
          "summary",
          "diversification_analysis",
          "strategic_advice",
          "top_risks",
          "top_opportunities",
          "detailed_report",
          "created_at"
        ],
        "title": "PortfolioAnalysisResponse"
      },
      "PortfolioCreate": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "avg_cost": {
            "type": "number",
            "title": "Avg Cost"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "quantity",
          "avg_cost"
        ],
        "title": "PortfolioCreate"
      },
...
      "PortfolioSummary": {
        "properties": {
          "total_market_value": {
            "type": "number",
            "title": "Total Market Value"
          },
          "total_unrealized_pl": {
            "type": "number",
            "title": "Total Unrealized Pl"
          },
          "total_pl_percent": {
            "type": "number",
            "title": "Total Pl Percent"
          },
          "day_change": {
            "type": "number",
            "title": "Day Change"
          },
          "holdings": {
            "items": {
              "$ref": "#/components/schemas/PortfolioItem"
            },
            "type": "array",
            "title": "Holdings"
          },
          "sector_exposure": {
            "items": {
              "$ref": "#/components/schemas/SectorExposure"
            },
            "type": "array",
            "title": "Sector Exposure"
          }
        },
        "type": "object",
        "required": [
          "total_market_value",
          "total_unrealized_pl",
          "total_pl_percent",
          "day_change",
          "holdings",
          "sector_exposure"
        ],
        "title": "PortfolioSummary"
      },
      "SearchResult": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "name"
        ],
        "title": "SearchResult"
      },
      "SectorExposure": {
        "properties": {
          "sector": {
            "type": "string",
            "title": "Sector"
          },
          "weight": {
            "type": "number",
            "title": "Weight"
          },
          "value": {
            "type": "number",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "sector",
          "weight",
          "value"
        ],
        "title": "SectorExposure"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}